data_acquisition/requirements.txt
pandas==
google-oauth
google-auth
google-api-core
google-api-python-client
geopy
jupyter